Password strength checker for view_site.php
A new button has been added to view_site.php to evaluate the strength of an existing password. A similar mechanism will be added to edit_site.php in a future update.
NOTE: I changed the database connection username to "test" again; please revert to "root" if you need to. This was on add_password, edit_site, view_pass, and view_site.