Admin Settings functionality complete
Created 'adminmenu.php', which serves as an admin console for password length and age requirements. The page updates the 'adminset' table in the database. This page is only accessible to users with admin privileges. NOTE: The initial values for PassAge and PassLength must be set in phpmyadmin, otherwise they will not update and will generate an error.

Updated 'index.php' to have an "Admin Settings" button that is only present to admin users. This button links to 'adminmenu.php'.

Updated 'view_pass.php' to show the alert icon when either the password age or length requirements are not met. These requirements are grabbed from the 'adminset' table. Only a single alert icon will appear per button.

Updated 'view_site.php' to display detailed alerts for password requirements. It provides a detailed message to notify the user of what must be updated. These requirements are also grabbed from the 'adminset' database.

Updated 'edit_site.php' to enforce password length requirements. If the user tries to enter a password under the admin's set password length, they will encounter an error and the password (along with any other edits) will not be updated.

Updated 'index.php' to have an "Admin Settings" button that is only present to admin users. This button links to 'adminmenu.php'.