Password Sharing notification feature
Added a one-time notification when a password is shared with another user. When the user who receives the shared password visits view_pass.php for the first time, text will indicate the name of the site(s) that have been shared with them. This is achieved by comparing the ViewerID and UserID of the currently logged in user. If the ShareScreen value/flag is 0, the user will be shown the notification. After, the value is changed to 1.
The styling of this notification could be improved.